Глава-6/requirements.txt
Django==3.1.14
djangorestframework==3.11.0
